+++ partiql-ast/src/main/kotlin/org/partiql/ast/sql/internal/InternalSqlDialect.kt
@@ -82,6 +82,13 @@ internal abstract class InternalSqlDialect : AstBaseVisitor<InternalSqlBlock, In
             t = t concat ")"
             t
         }
+        is Expr.BagOp -> {
+            var t = tail
+            t = t concat "("
+            t = visitExprBagOp(node, t)
+            t = t concat ")"
+            t
+        }
         else -> visitExpr(node, tail)
     }