+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/ir/Nodes.kt
@@ -54,7 +54,9 @@ import org.partiql.planner.internal.ir.builder.RexOpCaseBuilder
 import org.partiql.planner.internal.ir.builder.RexOpCoalesceBuilder
 import org.partiql.planner.internal.ir.builder.RexOpCollectionBuilder
 import org.partiql.planner.internal.ir.builder.RexOpErrBuilder
+import org.partiql.planner.internal.ir.builder.RexOpExceptBuilder
 import org.partiql.planner.internal.ir.builder.RexOpGlobalBuilder
+import org.partiql.planner.internal.ir.builder.RexOpIntersectBuilder
 import org.partiql.planner.internal.ir.builder.RexOpLitBuilder
 import org.partiql.planner.internal.ir.builder.RexOpNullifBuilder
 import org.partiql.planner.internal.ir.builder.RexOpPathIndexBuilder
@@ -66,6 +68,7 @@ import org.partiql.planner.internal.ir.builder.RexOpStructBuilder
 import org.partiql.planner.internal.ir.builder.RexOpStructFieldBuilder
 import org.partiql.planner.internal.ir.builder.RexOpSubqueryBuilder
 import org.partiql.planner.internal.ir.builder.RexOpTupleUnionBuilder
+import org.partiql.planner.internal.ir.builder.RexOpUnionBuilder
 import org.partiql.planner.internal.ir.builder.RexOpVarResolvedBuilder
 import org.partiql.planner.internal.ir.builder.RexOpVarUnresolvedBuilder
 import org.partiql.planner.internal.ir.builder.StatementQueryBuilder
@@ -322,6 +325,9 @@ internal data class Rex(
             is Subquery -> visitor.visitRexOpSubquery(this, ctx)
             is Select -> visitor.visitRexOpSelect(this, ctx)
             is TupleUnion -> visitor.visitRexOpTupleUnion(this, ctx)
+            is Union -> visitor.visitRexOpUnion(this, ctx)
+            is Intersect -> visitor.visitRexOpIntersect(this, ctx)
+            is Except -> visitor.visitRexOpExcept(this, ctx)
             is Err -> visitor.visitRexOpErr(this, ctx)
         }
 
@@ -744,6 +750,81 @@ internal data class Rex(
                 internal fun builder(): RexOpTupleUnionBuilder = RexOpTupleUnionBuilder()
             }
         }
+        
+        internal data class Union(
+          @JvmField
+          internal val setq: SetQuantifier,
+          @JvmField
+          internal val lhs: Rex,
+          @JvmField
+          internal val rhs: Rex,
+        ) : Op() {
+          internal override val children: List<PlanNode> by lazy {
+            val kids = mutableListOf<PlanNode?>()
+            kids.add(lhs)
+            kids.add(rhs)
+            kids.filterNotNull()
+          }
+    
+    
+          internal override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R =
+              visitor.visitRexOpUnion(this, ctx)
+    
+          internal companion object {
+            @JvmStatic
+            internal fun builder(): RexOpUnionBuilder = RexOpUnionBuilder()
+          }
+        }
+    
+        internal data class Intersect(
+          @JvmField
+          internal val setq: SetQuantifier,
+          @JvmField
+          internal val lhs: Rex,
+          @JvmField
+          internal val rhs: Rex,
+        ) : Op() {
+          internal override val children: List<PlanNode> by lazy {
+            val kids = mutableListOf<PlanNode?>()
+            kids.add(lhs)
+            kids.add(rhs)
+            kids.filterNotNull()
+          }
+    
+    
+          internal override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R =
+              visitor.visitRexOpIntersect(this, ctx)
+    
+          internal companion object {
+            @JvmStatic
+            internal fun builder(): RexOpIntersectBuilder = RexOpIntersectBuilder()
+          }
+        }
+    
+        internal data class Except(
+          @JvmField
+          internal val setq: SetQuantifier,
+          @JvmField
+          internal val lhs: Rex,
+          @JvmField
+          internal val rhs: Rex,
+        ) : Op() {
+          internal override val children: List<PlanNode> by lazy {
+            val kids = mutableListOf<PlanNode?>()
+            kids.add(lhs)
+            kids.add(rhs)
+            kids.filterNotNull()
+          }
+    
+    
+          internal override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R =
+              visitor.visitRexOpExcept(this, ctx)
+    
+          internal companion object {
+            @JvmStatic
+            internal fun builder(): RexOpExceptBuilder = RexOpExceptBuilder()
+          }
+        }
 
         internal data class Err(
             @JvmField internal val message: String,
@@ -952,60 +1033,78 @@ internal data class Rel(
         }
 
         internal data class Union(
-            @JvmField internal val lhs: Rel,
-            @JvmField internal val rhs: Rel,
+          @JvmField
+          internal val setq: SetQuantifier,
+          @JvmField
+          internal val lhs: Rel,
+          @JvmField
+          internal val rhs: Rel,
         ) : Op() {
-            override val children: List<PlanNode> by lazy {
-                val kids = mutableListOf<PlanNode?>()
-                kids.add(lhs)
-                kids.add(rhs)
-                kids.filterNotNull()
-            }
-
-            override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R = visitor.visitRelOpUnion(this, ctx)
-
-            internal companion object {
-                @JvmStatic
-                internal fun builder(): RelOpUnionBuilder = RelOpUnionBuilder()
-            }
+          internal override val children: List<PlanNode> by lazy {
+            val kids = mutableListOf<PlanNode?>()
+            kids.add(lhs)
+            kids.add(rhs)
+            kids.filterNotNull()
+          }
+    
+    
+          internal override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R =
+              visitor.visitRelOpUnion(this, ctx)
+    
+          internal companion object {
+            @JvmStatic
+            internal fun builder(): RelOpUnionBuilder = RelOpUnionBuilder()
+          }
         }
-
+    
         internal data class Intersect(
-            @JvmField internal val lhs: Rel,
-            @JvmField internal val rhs: Rel,
+          @JvmField
+          internal val setq: SetQuantifier,
+          @JvmField
+          internal val lhs: Rel,
+          @JvmField
+          internal val rhs: Rel,
         ) : Op() {
-            override val children: List<PlanNode> by lazy {
-                val kids = mutableListOf<PlanNode?>()
-                kids.add(lhs)
-                kids.add(rhs)
-                kids.filterNotNull()
-            }
-
-            override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R = visitor.visitRelOpIntersect(this, ctx)
-
-            internal companion object {
-                @JvmStatic
-                internal fun builder(): RelOpIntersectBuilder = RelOpIntersectBuilder()
-            }
+          internal override val children: List<PlanNode> by lazy {
+            val kids = mutableListOf<PlanNode?>()
+            kids.add(lhs)
+            kids.add(rhs)
+            kids.filterNotNull()
+          }
+    
+    
+          internal override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R =
+              visitor.visitRelOpIntersect(this, ctx)
+    
+          internal companion object {
+            @JvmStatic
+            internal fun builder(): RelOpIntersectBuilder = RelOpIntersectBuilder()
+          }
         }
-
+    
         internal data class Except(
-            @JvmField internal val lhs: Rel,
-            @JvmField internal val rhs: Rel,
+          @JvmField
+          internal val setq: SetQuantifier,
+          @JvmField
+          internal val lhs: Rel,
+          @JvmField
+          internal val rhs: Rel,
         ) : Op() {
-            override val children: List<PlanNode> by lazy {
-                val kids = mutableListOf<PlanNode?>()
-                kids.add(lhs)
-                kids.add(rhs)
-                kids.filterNotNull()
-            }
-
-            override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R = visitor.visitRelOpExcept(this, ctx)
-
-            internal companion object {
-                @JvmStatic
-                internal fun builder(): RelOpExceptBuilder = RelOpExceptBuilder()
-            }
+          internal override val children: List<PlanNode> by lazy {
+            val kids = mutableListOf<PlanNode?>()
+            kids.add(lhs)
+            kids.add(rhs)
+            kids.filterNotNull()
+          }
+    
+    
+          internal override fun <R, C> accept(visitor: PlanVisitor<R, C>, ctx: C): R =
+              visitor.visitRelOpExcept(this, ctx)
+    
+          internal companion object {
+            @JvmStatic
+            internal fun builder(): RelOpExceptBuilder = RelOpExceptBuilder()
+          }
         }
 
         internal data class Limit(
@@ -1280,4 +1379,9 @@ internal data class Rel(
         @JvmStatic
         internal fun builder(): RelBuilder = RelBuilder()
     }
+}
+
+internal enum class SetQuantifier {
+    ALL,
+    DISTINCT,
 }