+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/transforms/PlanTransform.kt
@@ -13,6 +13,7 @@ import org.partiql.planner.internal.ir.Identifier
 import org.partiql.planner.internal.ir.PartiQLPlan
 import org.partiql.planner.internal.ir.Rel
 import org.partiql.planner.internal.ir.Rex
+import org.partiql.planner.internal.ir.SetQuantifier
 import org.partiql.planner.internal.ir.Statement
 import org.partiql.planner.internal.ir.visitor.PlanBaseVisitor
 import org.partiql.planner.internal.utils.PlanUtils
@@ -236,6 +237,24 @@ internal object PlanTransform : PlanBaseVisitor<PlanNode, ProblemCallback>() {
         override fun visitRexOpTupleUnion(node: Rex.Op.TupleUnion, ctx: ProblemCallback) =
             org.partiql.plan.Rex.Op.TupleUnion(args = node.args.map { visitRex(it, ctx) })
 
+        override fun visitRexOpExcept(node: Rex.Op.Except, ctx: ProblemCallback) = org.partiql.plan.Rex.Op.Except(
+            lhs = visitRex(node.lhs, ctx),
+            rhs = visitRex(node.rhs, ctx),
+            setq = visitSetQuantifier(node.setq)
+        )
+
+        override fun visitRexOpIntersect(node: Rex.Op.Intersect, ctx: ProblemCallback) = org.partiql.plan.Rex.Op.Intersect(
+            lhs = visitRex(node.lhs, ctx),
+            rhs = visitRex(node.rhs, ctx),
+            setq = visitSetQuantifier(node.setq)
+        )
+
+        override fun visitRexOpUnion(node: Rex.Op.Union, ctx: ProblemCallback) = org.partiql.plan.Rex.Op.Union(
+            lhs = visitRex(node.lhs, ctx),
+            rhs = visitRex(node.rhs, ctx),
+            setq = visitSetQuantifier(node.setq)
+        )
+
         override fun visitRexOpErr(node: Rex.Op.Err, ctx: ProblemCallback) = org.partiql.plan.Rex.Op.Err(node.message)
 
         // RELATION OPERATORS
@@ -296,20 +315,28 @@ internal object PlanTransform : PlanBaseVisitor<PlanNode, ProblemCallback>() {
             }
         )
 
-        override fun visitRelOpUnion(node: Rel.Op.Union, ctx: ProblemCallback) = org.partiql.plan.Rel.Op.Union(
+        override fun visitRelOpExcept(node: Rel.Op.Except, ctx: ProblemCallback) = org.partiql.plan.Rel.Op.Except(
             lhs = visitRel(node.lhs, ctx),
             rhs = visitRel(node.rhs, ctx),
+            setq = visitSetQuantifier(node.setq)
         )
 
         override fun visitRelOpIntersect(node: Rel.Op.Intersect, ctx: ProblemCallback) = org.partiql.plan.Rel.Op.Intersect(
             lhs = visitRel(node.lhs, ctx),
             rhs = visitRel(node.rhs, ctx),
+            setq = visitSetQuantifier(node.setq)
         )
 
-        override fun visitRelOpExcept(node: Rel.Op.Except, ctx: ProblemCallback) = org.partiql.plan.Rel.Op.Except(
+        override fun visitRelOpUnion(node: Rel.Op.Union, ctx: ProblemCallback) = org.partiql.plan.Rel.Op.Union(
             lhs = visitRel(node.lhs, ctx),
             rhs = visitRel(node.rhs, ctx),
+            setq = visitSetQuantifier(node.setq)
         )
+
+        private fun visitSetQuantifier(node: SetQuantifier) = when (node) {
+            SetQuantifier.ALL -> org.partiql.plan.SetQuantifier.ALL
+            SetQuantifier.DISTINCT -> org.partiql.plan.SetQuantifier.DISTINCT
+        }
 
         override fun visitRelOpLimit(node: Rel.Op.Limit, ctx: ProblemCallback) = org.partiql.plan.Rel.Op.Limit(
             input = visitRel(node.input, ctx),